# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateWorkerRequestDto": {
        "properties": {
          "email": { "format": "email", "maxLength": 254, "type": "string" },
          "name": { "maxLength": 160, "minLength": 1, "type": "string" },
          "phone": { "maxLength": 40, "minLength": 1, "type": "string" },
          "plantId": { "format": "uuid", "type": "string" }
        },
        "required": ["plantId", "name"],
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": { "example": "AUTHORIZATION_DENIED", "type": "string" },
          "correlationId": { "format": "uuid", "type": "string" },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "items": { "type": "string" }, "type": "array" }
            ]
          },
          "path": {
            "example": "/api/v1/organizations/30000000-0000-4000-8000-000000000001/plants",
            "type": "string"
          },
          "statusCode": { "example": 403, "type": "integer" },
          "timestamp": { "format": "date-time", "type": "string" }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "path",
          "timestamp",
          "correlationId"
        ],
        "type": "object"
      },
      "MergeWorkerRequestDto": {
        "properties": {
          "canonicalWorkerId": { "format": "uuid", "type": "string" },
          "reason": { "maxLength": 300, "minLength": 1, "type": "string" }
        },
        "required": ["canonicalWorkerId"],
        "type": "object"
      },
      "Object": { "properties": {}, "type": "object" },
...
      "PageWorker": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/Worker" },
            "type": "array"
          },
          "page": { "minimum": 1, "type": "integer" },
          "pageSize": { "maximum": 100, "minimum": 1, "type": "integer" },
          "total": { "minimum": 0, "type": "integer" },
          "totalPages": { "minimum": 0, "type": "integer" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"],
        "type": "object"
      },
...
      "ResolveWorkerRequestDto": {
        "properties": {
          "reason": { "maxLength": 300, "minLength": 1, "type": "string" }
        },
        "type": "object"
      },
...
      "SetWorkerStateDto": {
        "properties": { "active": { "type": "boolean" } },
        "required": ["active"],
        "type": "object"
      },
...
      "Worker": {
        "properties": {
          "deactivatedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "email": { "format": "email", "nullable": true, "type": "string" },
          "id": { "format": "uuid", "type": "string" },
          "isActive": { "type": "boolean" },
          "name": { "example": "Trabajador ficticio", "type": "string" },
          "organizationId": { "format": "uuid", "type": "string" },
          "phone": { "nullable": true, "type": "string" },
          "plantId": { "format": "uuid", "type": "string" },
          "sourceRequestId": { "format": "uuid", "type": "string" },
          "status": {
            "enum": [
              "PROVISIONAL",
              "PROVISIONAL_VENCIDO",
              "ACTIVO",
              "RECHAZADO"
            ],
            "type": "string"
          },
          "statusChangedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "organizationId",
          "plantId",
          "sourceRequestId",
          "name",
          "status",
          "statusChangedAt",
          "isActive"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "supabase": {
        "bearerFormat": "Supabase access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Contrato backend del sistema de planta. Todos los ejemplos son ficticios.",
    "title": "Industrias Doradas API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/worker-requests": {
      "get": {
        "operationId": "WorkersController_listRequests",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "allOf": [{ "$ref": "#/components/schemas/Object" }],
              "default": 1,
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "allOf": [{ "$ref": "#/components/schemas/Object" }],
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "default": "active",
              "enum": ["all", "active", "inactive"],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": ["PENDING", "APPROVED", "REJECTED", "MERGED"],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plantId",
            "required": false,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Worker" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Workers"]
      },
      "post": {
        "operationId": "WorkersController_requestWorker",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkerRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Worker" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Workers"]
      }
    },
    "/api/v1/organizations/{organizationId}/worker-requests/{id}/approve": {
      "post": {
        "operationId": "WorkersController_approve",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Worker" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Workers"]
      }
    },
    "/api/v1/organizations/{organizationId}/worker-requests/{id}/merge": {
      "post": {
        "operationId": "WorkersController_merge",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MergeWorkerRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Worker" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Workers"]
      }
    },
    "/api/v1/organizations/{organizationId}/worker-requests/{id}/reject": {
      "post": {
        "operationId": "WorkersController_reject",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveWorkerRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Worker" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Workers"]
      }
    },
    "/api/v1/organizations/{organizationId}/workers": {
      "get": {
        "operationId": "WorkersController_listWorkers",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "allOf": [{ "$ref": "#/components/schemas/Object" }],
              "default": 1,
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "allOf": [{ "$ref": "#/components/schemas/Object" }],
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "default": "active",
              "enum": ["all", "active", "inactive"],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PROVISIONAL",
                "PROVISIONAL_VENCIDO",
                "ACTIVO",
                "RECHAZADO"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plantId",
            "required": false,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PageWorker" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Workers"]
      }
    },
    "/api/v1/organizations/{organizationId}/workers/{id}/state": {
      "patch": {
        "operationId": "WorkersController_setState",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SetWorkerStateDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Worker" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Workers"]
      }
    },
...
  },
  "servers": [],
...
}